feat: improve like handling
We'll use now a like counter in the post itself rather than doing
an expensive query every time.

Later on, we'll update this value when a new like is added
(and probably batch the count by using Redis / Kafka so that
posts with a lot of incoming likes don't get the row locked).